feat(extension) Use pre-build dynamic libraries.
By using pre-build dynamic libraries (when a new Wasmer version is
released), we simplify the build process of this project. Rust is no
longer required to build the PHP extension.
