+++ .gitignore
@@ -1,5 +1,4 @@
 /composer.lock
 /doc
-/extension/libwasmer_runtime_c_api.a
 /target/
 /vendor/
