--- extension/CREDITS
@@ -1,2 +1,0 @@
-wasm
-Ivan Enderlin
