--- extension/composer.json
@@ -1,12 +1,0 @@
-{
-    "name": "wasm",
-    "description": "Run WebAssembly binaries safely.",
-    "type": "extension",
-    "license": "BSD-3-Clause",
-    "authors": [
-        {
-            "name": "Ivan Enderlin",
-            "email": "ivan.enderlin@hoa-project.net"
-        }
-    ]
-}
