--- src/lib.rs
@@ -1,1 +1,0 @@
-// Empty file to use Cargo simply.
